third_party/WebKit/LayoutTests/fast/html/imports/import-same-url-expected.txt

PASS target1.import is non-null.
PASS target1.import is target2.import
PASS target1.import is target3.import
PASS successfullyParsed is true

TEST COMPLETE
